Modularisation du dashboard.
